Extract record death section into its own component
